fix: update broken layout becuase DS changes
